maxQueryWordsForQueryType moved to conf/validations
